+++ wsidicomizer/sources/tiffslide/tiffslide_source.py
@@ -130,7 +130,7 @@ class TiffSlideSource(OpenSlideLikeSource):
             self._blank_color,
             self._base_level_offset,
             self._base_level_size,
-            self.metadata.image,
+            self.metadata.pyramid.image,
             level_index,
             self._tile_size,
             self._encoder,